+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version 0.30.1:
+- ListGenerator extends Iterable
+- Fixed a bug in table SAS generation where table names with capital letters were not signed correctly.
+
 ## Version 0.30.0:
 
 ### All: